(김동우) edit: 회원가입 페이지 리팩토링
